Modified the Tower class and Item class, shop is now functional, but items aren't able to be bought and put into players inventory yet.

+++ src/main/java/seng201/team0/models/DairyTower.java
@@ -6,7 +6,7 @@ import java.util.Random;
 public class DairyTower extends Tower {
     private static final List<String> dairyNames = List.of("Milk Tower", "Cheese Tower", "Yogurt Tower", "Butter Tower");
 
-    public DairyTower(float towerSpeed, int towerFillAmount, int buyPrice, int gameBonus) {
+    public DairyTower(int towerSpeed, int towerFillAmount, int buyPrice, int gameBonus) {
         super(towerSpeed, towerFillAmount, buyPrice, gameBonus);
     }
